Added cronjob for auto closing overdue polls
Cronjob runs once every minute, which essentially means that as a worst case scenario a user can vote on a poll 59seconds after its supposed to be closed. Can add a check in /api/vote if the poll endtime is less than current time as well if we deem this necessary.