GetServedPolls endpoint, change serve poll
Added an endpoint for getting all polls served to an IoT device, GET /api/iot/servedPolls/{deviceId}.
Added logic to the existing endpoint for serving polls to check if the poll is closed or private, we now only allow open and public polls to be served to IoT devices. Changed the structure of this endpoint from having IoT device id in api path and pollid in request body to making the request go to /api/IoT/{IoTId}/{pollId} with an empty request body.