Handle errors on invalid IoT Guid, design changes
Added logic to handle invalid response on selecting IoT by id. Changed the UI to be more intuitive (green button on right side, red on left) as well as making some big changes to the overall design of the page.

Todo:
Need to handle errors on voting as well as figure out how to handle when the queue is empty, for now i avoided error by simply returning if this is the case, see NextPollButton.tsx. Note that navigating to / works, so it takes you back to the start screen, however /id does not seem to do anything.
In terms of backend I will add a check tomorrow to see if a vote is performed by an IoT device and if so remove it from the IoT poll queue. I will also add functionality to closePoll so that when a poll closes it is removed from all IoT queues OR remove a poll from IoT queue if an IoT device attempts to vote on a closed poll.