Added .env, fixed issue with logging in
Added .env for development (.env) and production (.env.production) and replaced hardcoded paths with the new variables. Added open in new window to API button. Fixed an issue where attempting to log in with invalid credentials set an unknown value for user in localstorage, causing the whole site to malfunction.